requirements.txt
# Core dependencies
pandas==2.0.1
streamlit==1.25.0
python-dotenv==1.0.0

# Web crawling
scrapy==2.8.0
playwright==1.39.0
requests==2.30.0

# YouTube API
youtube-transcript-api==0.6.1

# OpenAI integration
openai==0.27.8

# Notifications
slack-sdk==3.21.3

# Utils
tqdm==4.65.0
beautifulsoup4==4.12.2 